requirements.txt
adal==1.2.7
adlfs==2022.10.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
appnope==0.1.3
apprise==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==1.0.1
asn1crypto==1.5.1
asttokens==2.0.8
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
azure-core==1.25.1
azure-datalake-store==0.0.52
azure-identity==1.11.0
azure-storage-blob==12.13.1
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
certifi==2022.6.15.1
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
cloudpickle==2.2.0
commonmark==0.9.1
coolname==1.1.0
crashtest==0.3.1
croniter==1.3.7
cryptography==36.0.2
cycler==0.11.0
dask==2022.9.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distributed==2022.9.1
docker==6.0.0
dulwich==0.20.46
entrypoints==0.4
executing==1.0.0
fastapi==0.82.0
fastjsonschema==2.16.1
filelock==3.8.0
flake8==5.0.4
fonttools==4.37.2
frozenlist==1.3.1
fsspec==2022.8.2
google-auth==2.11.0
greenlet==1.1.3
griffe==0.21.0
h11==0.12.0
HeapDict==1.0.1
html5lib==1.1
httpcore==0.15.0
httpx==0.23.0
identify==2.5.5
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
isodate==0.6.1
jaraco.classes==3.2.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.16.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyterlab==3.4.6
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.1
keyring==23.9.1
kiwisolver==1.4.4
kubernetes==24.2.0
locket==1.0.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
more-itertools==8.14.0
msal==1.19.0
msal-extensions==1.0.0
msgpack==1.0.4
msrest==0.7.1
multidict==6.0.2
multitasking==0.0.11
mypy-extensions==0.4.3
Nasdaq-Data-Link==1.0.4
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
nodeenv==1.7.0
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
oauthlib==3.2.1
OpenPermID==0.5
orjson==3.8.0
oscrypto==1.3.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.10.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
portalocker==2.5.1
pre-commit==2.20.0
prefect==2.4.0
prefect-azure==0.2.0
prefect-dask==0.2.0.post1
prefect-snowflake==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.4.0
pylev==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.2.1
readchar==4.0.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
Send2Trash==1.8.0
shellingham==1.5.0
six==1.16.0
slack-sdk==3.18.3
sniffio==1.3.0
snowflake-connector-python==2.7.12
snowflake-sqlalchemy==1.4.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.0
starlette==0.19.1
tblib==1.7.0
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
toolz==0.12.0
tornado==6.1
traitlets==5.4.0
typer==0.6.1
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
virtualenv==20.16.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
xattr==0.9.9
yarl==1.8.1
yfinance==0.1.74
zict==2.2.0
zipp==3.8.1
